+++ src/tatc/schemas/orbit/__init__.py
@@ -5,7 +5,7 @@ Object schemas for orbits.
 """
 
 from .circular import CircularOrbit
-from .geostationary import GeostationaryOrbit
+from .geosynchronous import GeosynchronousOrbit
 from .gp import GeneralPerturbationsOrbit
 from .keplerian import KeplerianOrbit
 from .molniya import MolniyaOrbit
@@ -15,7 +15,7 @@ from .tundra import TundraOrbit
 AllOrbits = (
     CircularOrbit
     | GeneralPerturbationsOrbit
-    | GeostationaryOrbit
+    | GeosynchronousOrbit
     | KeplerianOrbit
     | MolniyaOrbit
     | SunSynchronousOrbit
@@ -26,7 +26,7 @@ __all__ = [
     "AllOrbits",
     "CircularOrbit",
     "GeneralPerturbationsOrbit",
-    "GeostationaryOrbit",
+    "GeosynchronousOrbit",
     "KeplerianOrbit",
     "MolniyaOrbit",
     "SunSynchronousOrbit",